requirements.txt
dicttoxml==1.7.4
numpy==1.21.2
pandas==1.3.3
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
toml==0.10.2
xmltodict==0.12.0
